bot/requirements.txt
dnspython==2.3.0
python-whois==0.8.0
requests==2.31.0
beautifulsoup4==4.12.2
jinja2==3.1.2
